Remove libcola's circular libtopology dependency.
libtopology and libcola both depended on code in each other.  This
change takes the topology code in libcola and refactors it into a addon
in libtopology.  This addon implements an interface in libcola that can
be used to replace some of the libcola functionality and provide
topology preserving layout.

This can be achieved with:
	cola::ConstrainedFDLayout *layout = ...
	topology::ColaTopologyAddon topology(topologyNodes, topologyEdges);
        layout->setTopology(&topology);