Speed up libvpsc by 50% on some large problems.
libvpsc used to keep a blocks list as a STL set.  It would do a lot of
operations -- like iterating over the set and inserting and removing
objects -- that had unnecessarily bad time complexity.  The previous
implementation of Blocks::cleanup() was especially bad.  With this change
the blocks list is kept as a vector.  Iterating over it is O(n), additions
are O(1) and the cleanup is O(n).