Fix a bug with pins on outside edge of visibility graph.
Correct visibility for pins or connector endpoints on the leading or
trailing edge of the visibility graph which may only have visibility in
the outward direction where there will not be a possible path.  We fix
this by giving them visibility in both directions along the leading or
trailing edge.