Improve speed of A* routing step on big problem instances.
Rather than checking the entire PENDING list to see if we have
encountered a node before, we now have the vertices keep a list of
pointers to their own pending nodes in the search.  Thus we can now do
this check in almost constant time.  This happens very frequently in the
inner loop of the search and speeds up the routing on large examples
by nearly 2X.