Adapt tests
Having `listAllPossibleTreePlacements()` now return ordinal placements, even when cardinal placements are
preferred, results in some necessary adjustments to our tests.

In the case of `nudgeopt.cpp`, the adjustment is due to instability in the sorting methods employed at the start
of the `chooseBestPlacement()` function. `S` placement is now chosen instead of `W`, due to arbitrary ordering.