Repair `EdgePoint` pruning in `libtopology`
To construct a `BendConstraint` at an `EdgePoint` requires that
at least one of the two incident segments have non-zero length
in the conjugate dimension.

We add a check in `EdgePoint::createBendConstraint()` to skip
the case where both of these lengths are zero.

Such a case can arise when pruning degenerate EdgePoints, four
or more of which are aligned in the scanning dimension.